refactor: use TacoSigner instead of ethers.Signer and remove provider dependency from TacoSigner interface and implementations
