+++ packages/shared/src/types.ts
@@ -1,5 +1,6 @@
 import { ethers } from 'ethers';
 
+import { TacoSigner } from './taco-interfaces';
 import { Account, PublicClient } from './viem/types';
 
 export type ChecksumAddress = `0x${string}`;
@@ -7,4 +8,5 @@ export type HexEncodedBytes = string;
 export type Base64EncodedBytes = string;
 
 export type ProviderLike = ethers.providers.Provider | PublicClient;
-export type SignerLike = ethers.Signer | Account;
+
+export type SignerLike = TacoSigner | Account;
