+++ packages/shared/src/viem/index.ts
@@ -1,3 +1,4 @@
-export * from './ethers-viem-utils';
+export * from './taco-provider';
+export * from './taco-signer';
 export * from './type-guards';
 export * from './types';
